requirements.txt
uvicorn==0.20.0
PyMySQL==1.0.2
mysql-connector-python==8.0.31
fastapi==0.88.0
aiomysql==0.1.1
pydantic==1.10.2
SQLAlchemy==1.4.41
pandas==1.5.2
opencv-python==4.6.0.66
keras==2.11.0
scikit-learn==1.1.3
numpy==1.23.2
databases==0.6.2
cryptography
python-dotenv==0.20.0
sqlalchemy_utils==0.38.3
python-multipart